no-bitcoind: domain reorganization and guidelines
