+++ requirements.txt
@@ -4,3 +4,4 @@ git+https://github.com/Conio/pybitcointools.git
 retrying
 connectrum
 async-timeout
+aiohttp
