+++ spruned/abstracts.py
