+++ spruned/services/bitcoind_rpc_service.py
