+++ spruned/services/bitpay_service.py
@@ -1,20 +1,21 @@
 from spruned import settings
-from spruned.service.abstract import RPCAPIService
-from spruned.third_party.http_client import HTTPClient
+from spruned.abstracts import RPCAPIService
+from spruned.services.http_client import HTTPClient
 
 
-class BlockexplorerService(RPCAPIService):
+class BitpayService(RPCAPIService):
     def __init__(self, coin, httpclient=HTTPClient):
         assert coin == settings.Network.BITCOIN
-        self.client = httpclient(baseurl='https://blockexplorer.com/api/')
+        self.client = httpclient(baseurl='https://insight.bitpay.com/api/')
 
     def getrawtransaction(self, txid, **_):
         data = self.client.get('tx/' + txid)
         return data and {
             'rawtx': None,
             'blockhash': data['blockhash'],
+            'size': None,
             'txid': txid,
-            'source': 'blockexplorer.com'
+            'source': 'insight.bitpay.com'
         }
 
     def getblock(self, blockhash):
