+++ spruned/services/blockexplorer_service.py
@@ -1,21 +1,20 @@
 from spruned import settings
-from spruned.service.abstract import RPCAPIService
-from spruned.third_party.http_client import HTTPClient
+from spruned.abstracts import RPCAPIService
+from spruned.services.http_client import HTTPClient
 
 
-class BitpayService(RPCAPIService):
+class BlockexplorerService(RPCAPIService):
     def __init__(self, coin, httpclient=HTTPClient):
         assert coin == settings.Network.BITCOIN
-        self.client = httpclient(baseurl='https://insight.bitpay.com/api/')
+        self.client = httpclient(baseurl='https://blockexplorer.com/api/')
 
     def getrawtransaction(self, txid, **_):
         data = self.client.get('tx/' + txid)
         return data and {
             'rawtx': None,
             'blockhash': data['blockhash'],
-            'size': None,
             'txid': txid,
-            'source': 'insight.bitpay.com'
+            'source': 'blockexplorer.com'
         }
 
     def getblock(self, blockhash):
